Revert naming menu to single implementation
No need to make use of AnvilGUI as an alternative if I have more control
over the update process of IF now.
